+++ boreas/boreas_io.c
@@ -240,7 +240,7 @@ put_host_on_queue (kb_t kb, char *addr_str)
 }
 
 /**
- * @brief Checks if the finish signal is already set as last item in the queue.
+ * @brief Checks if the finish signal is already set.
  *
  * @param main_kb  kb to use
  * @return 1 if it is already set. 0 otherwise.
@@ -248,18 +248,30 @@ put_host_on_queue (kb_t kb, char *addr_str)
 int
 finish_signal_on_queue (kb_t main_kb)
 {
-  struct kb_item *last_queue_item;
-  int ret;
+  static gboolean fin_msg_already_on_queue = FALSE;
+  struct kb_item *queue_items = NULL;
+  int ret = 0;
 
-  ret = 0;
-  last_queue_item =
-    kb_item_get_single (main_kb, ALIVE_DETECTION_QUEUE, KB_TYPE_STR);
+  if (fin_msg_already_on_queue)
+    return 1;
 
-  if (last_queue_item && (last_queue_item->type == KB_TYPE_STR)
-      && (!g_strcmp0 (last_queue_item->v_str, ALIVE_DETECTION_FINISHED)))
-    ret = 1;
-
-  kb_item_free (last_queue_item);
+  /* Check if it was already set throught the whole items under the key.
+     If so, set the static variable to avoid quering redis unnecessarily. */
+  queue_items =
+    kb_item_get_all (main_kb, ALIVE_DETECTION_QUEUE);
+  if (queue_items)
+    {
+      while (queue_items)
+        {
+          if (!g_strcmp0 (queue_items->v_str, ALIVE_DETECTION_FINISHED))
+            {
+              fin_msg_already_on_queue = TRUE;
+              ret = 1;
+            }
+          queue_items = queue_items->next;
+        }
+      kb_item_free (queue_items);
+    }
   return ret;
 }